remove max_filesize in html, check for imagick before make thumb, autosave script updated for DB

+++ inc/display_file.php
@@ -41,19 +41,22 @@ if($count > 0){
         ?>
             <a class='align_right' href='delete_file.php?id=<?php echo $uploads_data['id'];?>&type=<?php echo $uploads_data['type'];?>&item_id=<?php echo $uploads_data['item_id'];?>' onClick="return confirm('Delete this file ?');"><img src='themes/<?php echo $_SESSION['prefs']['theme'];?>/img/trash.png' title='delete' alt='delete' /></a>
         <?php
-        // Get file extension to display thumbnail if it's an image
-        $ext = get_ext($uploads_data['real_name']);
-        if ($ext === 'jpg' || $ext === 'jpeg' || $ext === 'JPG' || $ext === 'png' || $ext === 'gif'){
-            $filepath = 'uploads/'.$uploads_data['long_name'];
-            $filesize = filesize('uploads/'.$uploads_data['long_name']);
-            $thumbpath = 'uploads/'.$uploads_data['long_name'].'_th.'.$ext;
-            // Make thumbnail only if it isn't done already and if size < 2 Mbytes
-            if(!file_exists($thumbpath) && $filesize <= 2000000){
-                make_thumb($filepath,$ext,$thumbpath,150);
+        // Check if ImageMagick is dispo
+        if (class_exists('Imagick')) {
+            // Get file extension to display thumbnail if it's an image
+            $ext = get_ext($uploads_data['real_name']);
+            if ($ext === 'jpg' || $ext === 'jpeg' || $ext === 'JPG' || $ext === 'png' || $ext === 'gif'){
+                $filepath = 'uploads/'.$uploads_data['long_name'];
+                $filesize = filesize('uploads/'.$uploads_data['long_name']);
+                $thumbpath = 'uploads/'.$uploads_data['long_name'].'_th.'.$ext;
+                // Make thumbnail only if it isn't done already and if size < 2 Mbytes
+                if(!file_exists($thumbpath) && $filesize <= 2000000){
+                    make_thumb($filepath,$ext,$thumbpath,150);
+                }
+                echo "<div class='center'>";
+                echo "<img src='".$thumbpath."' alt='' /></div>";
             }
-            echo "<div class='center'>";
-            echo "<img src='".$thumbpath."' alt='' /></div>";
-        }
+        } //end thumb
         echo "<img src='themes/".$_SESSION['prefs']['theme']."/img/attached_file.png' alt='' /> <a href='download.php?id=".$uploads_data['id']."&f=".$uploads_data['long_name']."&name=".$uploads_data['real_name']."' target='_blank'>".$uploads_data['real_name']."</a>
         <span class='filesize'> (".format_bytes(filesize('uploads/'.$uploads_data['long_name'])).")</span><br />";
         echo "<img src='themes/".$_SESSION['prefs']['theme']."/img/comments.png' alt='comment' /> <p class='editable' id='comment_".$uploads_data['id']."'>".stripslashes($uploads_data['comment'])."</p></div>";
